+++ packages/cubejs-query-orchestrator/src/orchestrator/PreAggregationLoadCache.ts
@@ -191,32 +191,25 @@ export class PreAggregationLoadCache {
   }
 
   public async keyQueryResult(sqlQuery: QueryWithParams, waitForRenew: boolean, priority: QueuePriority) {
-    const [query, values, queryOptions] = sqlQuery;
+    const memoKey = this.queryCache.refreshKeyCacheKey(sqlQuery, this.dataSource);
 
-    if (!this.queryResults[this.queryCache.queryRedisKey([query, values])]) {
-      this.queryResults[this.queryCache.queryRedisKey([query, values])] = await this.queryCache.cacheQueryResult(
-        query,
-        values,
-        [query, values],
+    if (!this.queryResults[memoKey]) {
+      this.queryResults[memoKey] = await this.queryCache.cacheRefreshKeyResult(
+        sqlQuery,
         60 * 60,
         {
-          renewalThreshold: this.queryCache.options.refreshKeyRenewalThreshold
-            || queryOptions?.renewalThreshold || 2 * 60,
-          renewalKey: [query, values],
           waitForRenew,
           priority,
           requestId: this.requestId,
           dataSource: this.dataSource,
-          useInMemory: true,
-          external: queryOptions?.external
         }
       );
     }
-    return this.queryResults[this.queryCache.queryRedisKey([query, values])];
+    return this.queryResults[memoKey];
   }
 
-  public hasKeyQueryResult(keyQuery) {
-    return !!this.queryResults[this.queryCache.queryRedisKey(keyQuery)];
+  public hasKeyQueryResult(keyQuery: QueryWithParams) {
+    return !!this.queryResults[this.queryCache.refreshKeyCacheKey(keyQuery, this.dataSource)];
   }
 
   public async getQueryStage(stageQueryKey) {